+++ examples/io/viewers/viewer3D-6-clipping.cpp
@@ -27,6 +27,15 @@
  * This file is part of the DGtal library.
  */
 
+
+/**
+ * Example of visualization after adding two clipping planes
+ * @see \ref DGtalGLV_CLipping
+ * \image html  visuClipping2.png "Visualization after adding a second clipping plane (1,0,0,-4.9) ."
+ * \example io/viewers/viewer3D-6-clipping.cpp
+ */
+
+
 ///////////////////////////////////////////////////////////////////////////////
 #include <iostream>