+++ examples/io/viewers/viewer3D-7bis-stdplane.cpp
@@ -27,6 +27,25 @@
  * This file is part of the DGtal library.
  */
 
+/**
+   This example shows the recognition of a simple standard digital
+   plane with the Chord algorithm (class
+   ChordGenericStandardPlaneComputer ). Green points belong to the naive
+   plane. Grey points show farther points that also belong to the
+   strip.
+
+@see \ref modulePlaneRecognition_sec26
+
+@verbatim
+$ ./examples/io/viewers/viewer3D-7bis-stdplane
+@endverbatim
+
+@image html coba-stdplane.png "Piece of standard plane."
+@image latex coba-stdplane.png "Piece of standard plane." width=5cm
+
+\example io/viewers/viewer3D-7bis-stdplane.cpp
+*/
+
 ///////////////////////////////////////////////////////////////////////////////
 #include <cstdlib>
 #include <iostream>